Rename package to pkitesting to match the module name
