+++ pkitesting/src/test/java/io/netty/pkitesting/der/DerWriterTest.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package io.netty.testcert.der;
+package io.netty.pkitesting.der;
 
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
